Refactoring to use Java backed code to invoke Native SOLR Admin Console
