--- repository/src/main/java/org/orderofthebee/addons/support/tools/repo/search/SolrAdminConsole.java
@@ -1,7 +1,0 @@
-package org.orderofthebee.addons.support.tools.repo.search;
-
-public interface SolrAdminConsole {
-
-    int getCascadeTrackerPendingCount(String coreName);
-
-}
