+++ repository/src/main/resources/alfresco/subsystems/Search/solr6/ootbee-support-tools-context.xml
@@ -4,9 +4,10 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
         http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="solrAdminConsole" class="org.orderofthebee.addons.support.tools.repo.search.solr6.SolrAdminConsoleImpl"
-          init-method="init">
-        <property name="solrHttpClientFactory" ref="solrHttpClientFactory" />
+    <bean id="solrAdminNativeClient" class="org.orderofthebee.addons.support.tools.repo.search.solr6.SolrAdminNativeClientImpl">
+        <property name="httpClient">
+            <bean factory-bean="solrHttpClientFactory" factory-method="getHttpClient" />
+        </property>
         <property name="baseUrl" value="${solr.baseUrl}" />
     </bean>
